+++ azcam/azcam-mods/releases.md
@@ -6,6 +6,12 @@
 
 ### Version 1.1.6 - 2025 Dec 10
  * Set `exposure.shutter_delay = 500` msec on testing with 1Kx1K images acquired with MODS2R camera and the pinhole sieve mask, 1s exposure. Stop seeing streaks at 300ms.  The old controller used 1s, which was arbitrary.
+ * `detector_mods.py` updated, now includes 32 overscan columns from Mike Lesser
+ * `console.py` - new console setup method for azcam-mods from Mike Lesser
+ * `server.py` - modified server setup, with changes for the new Archon control files
+
+See the `azcam/server/MODS*/` folders for the updated Archon control files (`.ncf`) and changes in filenames
+for some elements to support the azcam-console.
 
 ### Version 1.1.5 - 2025 Nov 21
  * Added `exposure.shutter_delay = 250` to add a 250-msec shutter delay before starting readout.  This is needed because of the slower MODS shutter.  This is an initial value that will be optimized later to be a best for all 4 cameras.  Change was introduced by Mike Lesser during MODS detector testing.